requirements.txt
cloudinary
django
django-cors-headers
django-filter
djangorestframework
djangorestframework_simplejwt
djoser
factory-boy
gunicorn
mongoengine
numpy
pandas
Pillow
pre-commit
psycopg2-binary
python-dotenv
rake-nltk
requests
scikit-learn
social-auth-app-django
translate
unidecode
